feat: 셀 내부 데이터 분리 (#4)
- 셀 내부의 코드 각각 폴더링 후 정리
