+++ Presentation/Cell/ITunesThumbnailCell.swift
@@ -7,7 +7,6 @@
 
 import UIKit
 import Domain
-
 internal import SnapKit
 internal import Then
 
@@ -15,10 +14,11 @@ final class ITunesThumbnailCell: UICollectionViewCell {
     
     static let identifier = "ITunesThumbnailCell"
 
-    private let storeButton = AppStoreStyleButton()
+    // 블러처리된 뷰 및 그라데이션 뷰
     private let blurView = UIVisualEffectView(effect: UIBlurEffect(style: .systemUltraThinMaterial))
     private let gradientView = UIView()
     
+    /// 그라데이션 설정
     private lazy var gradientLayer = CAGradientLayer().then {
         $0.frame = self.contentView.bounds
         $0.colors = [
@@ -28,25 +28,37 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         $0.locations = [0.5, 1.0]
     }
     
-    private lazy var imageTtileStackView = UIStackView().then {
+    /// 프로필 제목 + 작가 + 연도 스택
+    private lazy var imageTitleStackView = UIStackView().then {
         $0.axis = .vertical
         $0.spacing = 4
         $0.addArrangedSubview(imageTitleLabel)
         $0.addArrangedSubview(imageSubtitleLabel)
     }
     
+    /// 받기 + 앱 내 구입 스택
+    private lazy var imageButtonStackView = UIStackView().then {
+        $0.axis = .vertical
+        $0.spacing = 4
+        $0.addArrangedSubview(downLoadLabel)
+        $0.addArrangedSubview(purchaseLabel)
+    }
+    
+    /// 썸네일
     private let imageView = UIImageView().then {
         $0.contentMode = .scaleAspectFill
         $0.layer.cornerRadius = 8
         $0.clipsToBounds = true
     }
     
+    /// 작은 이미지
     private let profileView = UIImageView().then {
         $0.contentMode = .scaleAspectFill
         $0.layer.cornerRadius = 16
         $0.clipsToBounds = true
     }
 
+    /// 타이틀 - 작품명
     private let titleLabel = UILabel().then {
         $0.font = .systemFont(ofSize: 20, weight: .semibold)
         $0.textColor = .label
@@ -54,6 +66,7 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         $0.clipsToBounds = true
     }
 
+    /// 서브 타이틀 - 아티스트 + 연도
     private let subtitleLabel = UILabel().then {
         $0.font = .systemFont(ofSize: 16, weight: .regular)
         $0.textColor = .secondaryLabel
@@ -61,6 +74,7 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         $0.clipsToBounds = true
     }
     
+    /// 이미지 내부 타이틀
     private let imageTitleLabel = UILabel().then {
         $0.font = .systemFont(ofSize: 16, weight: .semibold)
         $0.textColor = .systemBackground
@@ -68,6 +82,7 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         $0.clipsToBounds = true
     }
 
+    /// 이미지 내부 서브 타이틀
     private let imageSubtitleLabel = UILabel().then {
         $0.font = .systemFont(ofSize: 14, weight: .regular)
         $0.textColor = .white
@@ -75,15 +90,36 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         $0.clipsToBounds = true
     }
 
+    /// 장르
     private let genreLabel = UILabel().then {
         $0.font = .systemFont(ofSize: 14, weight: .medium)
         $0.textColor = .systemBlue
     }
 
+    /// 가격
     private let priceLabel = UILabel().then {
         $0.font = .systemFont(ofSize: 16, weight: .medium)
         $0.textColor = .systemGray
         $0.clipsToBounds = true
+    }
+    
+    /// 이미지 내부 받기 버튼
+    private let downLoadLabel = PaddedLabel(vertical: 8, horizontal: 24).then {
+        $0.text = "받기"
+        $0.font = .systemFont(ofSize: 16, weight: .bold)
+        $0.textColor = .white
+        $0.textAlignment = .center
+        $0.backgroundColor = UIColor.lightGray.withAlphaComponent(0.5)
+        $0.layer.cornerRadius = 16
+        $0.clipsToBounds = true
+    }
+
+    /// 이미지 내부 구입
+    private let purchaseLabel = UILabel().then {
+        $0.text = "앱 내 구입"
+        $0.font = .systemFont(ofSize: 12, weight: .regular)
+        $0.textColor = .white
+        $0.textAlignment = .center
     }
     
     override init(frame: CGRect) {
@@ -101,14 +137,16 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         super.prepareForReuse()
     }
     
+    /// 뷰 추가 및 설정
     private func configureViews() {
         [titleLabel, subtitleLabel, imageView, genreLabel, priceLabel]
             .forEach { contentView.addSubview($0) }
-        [blurView, gradientView, imageTtileStackView, profileView, storeButton]
+        [blurView, gradientView, imageTitleStackView, profileView, imageButtonStackView]
             .forEach { imageView.addSubview($0) }
         gradientView.layer.addSublayer(gradientLayer)
     }
 
+    /// 오토레이아웃 설정
     private func configureLayout() {
 
         titleLabel.snp.makeConstraints {
@@ -146,26 +184,26 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         }
     }
     
+    /// 이미지 내부 오토레이아웃 설정
     private func configureImageLayout() {
         profileView.snp.makeConstraints {
-            $0.leading.equalToSuperview().inset(16)
-            $0.centerY.equalTo(storeButton).offset(8)
+            $0.leading.bottom.equalToSuperview().inset(16)
             $0.size.equalTo(CGSize(width: 64, height: 64))
         }
         
-        imageTtileStackView.snp.makeConstraints {
+        imageTitleStackView.snp.makeConstraints {
             $0.leading.equalTo(profileView.snp.trailing).offset(8)
-            $0.centerY.equalTo(storeButton).offset(8)
+            $0.centerY.equalTo(profileView).offset(8)
             $0.width.equalTo(128)
         }
 
-        storeButton.snp.makeConstraints {
-            $0.trailing.equalToSuperview().inset(8)
-            $0.bottom.equalToSuperview().inset(32)
-            $0.height.equalTo(48)
+        imageButtonStackView.snp.makeConstraints {
+            $0.trailing.equalToSuperview().inset(16)
+            $0.centerY.equalTo(profileView).offset(8)
         }
     }
 
+    /// 셀 외부에서 데이터 업데이트
     func configure(with item: ITunes) {
         imageView.load(url: item.imageURL)
         profileView.load(url: item.imageURL)
@@ -177,92 +215,8 @@ final class ITunesThumbnailCell: UICollectionViewCell {
         priceLabel.text = item.priceText
     }
     
+    /// 접근
     var getImageView: UIImageView {
         imageView
-    }
-}
-
-// MARK: - UIImageView URL load helper (임시)
-extension UIImageView {
-    func load(url: URL) {
-        DispatchQueue.global().async {
-            if let data = try? Data(contentsOf: url),
-               let image = UIImage(data: data) {
-                DispatchQueue.main.async {
-                    self.image = image
-                }
-            }
-        }
-    }
-}
-
-
-
-// MARK: - App Store Style Button
-final class AppStoreStyleButton: UIView {
-    private let titleLabel = PaddedLabel(vertical: 8, horizontal: 24).then {
-        $0.text = "받기"
-        $0.font = .systemFont(ofSize: 16, weight: .bold)
-        $0.textColor = .white
-        $0.textAlignment = .center
-        $0.backgroundColor = UIColor.lightGray.withAlphaComponent(0.5)
-        $0.layer.cornerRadius = 16
-        $0.clipsToBounds = true
-    }
-
-    private let subtitleLabel = UILabel().then {
-        $0.text = "앱 내 구입"
-        $0.font = .systemFont(ofSize: 12, weight: .regular)
-        $0.textColor = .white
-        $0.textAlignment = .center
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private func setupView() {
-        
-        [titleLabel, subtitleLabel]
-            .forEach { addSubview($0) }
-
-        titleLabel.snp.makeConstraints {
-            $0.top.horizontalEdges.equalToSuperview().inset(8)
-        }
-        
-        subtitleLabel.snp.makeConstraints {
-            $0.centerX.equalTo(titleLabel.snp.centerX)
-            $0.top.equalTo(titleLabel.snp.bottom).offset(8)
-        }
-    }
-}
-
-// MARK: - PaddedLabel
-final class PaddedLabel: UILabel {
-    
-    var padding: UIEdgeInsets
-
-    init(vertical: CGFloat, horizontal: CGFloat) {
-        padding = UIEdgeInsets(top: vertical, left: horizontal, bottom: vertical, right: horizontal)
-        super.init(frame: .zero)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func drawText(in rect: CGRect) {
-        super.drawText(in: rect.inset(by: padding))
-    }
-
-    override var intrinsicContentSize: CGSize {
-        let size = super.intrinsicContentSize
-        return CGSize(width: size.width + padding.left + padding.right,
-                      height: size.height + padding.top + padding.bottom)
     }
 }
